+++ packages/uiengine-theme/src/vue/router/index.js
@@ -30,11 +30,6 @@ const router = new Router({
       component: mainComponent('search')
     },
     {
-      name: 'sketch',
-      path: '/_sketch/',
-      component: mainComponent('sketch')
-    },
-    {
       name: 'settings',
       path: '/_settings/',
       component: mainComponent('settings')